backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
openai==1.3.7
faiss-cpu==1.7.4
numpy==1.24.3
sqlalchemy==2.0.23
sentence-transformers==2.2.2
wikipedia==1.4.0
networkx==3.2.1
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-multipart==0.0.6
pydantic-settings==2.1.0

